+++ java/veridot-core/src/main/java/io/github/cyfko/veridot/core/exceptions/BrokerTransportException.java
@@ -1,6 +1,6 @@
 package io.github.cyfko.veridot.core.exceptions;
 
-public class BrokerTransportException extends RuntimeException {
+public class BrokerTransportException extends VeridotException {
     public BrokerTransportException(String message) {
         super(message);
     }