+++ java/veridot-core/src/main/java/io/github/cyfko/veridot/core/exceptions/DataDeserializationException.java
@@ -1,6 +1,6 @@
 package io.github.cyfko.veridot.core.exceptions;
 
-public class DataDeserializationException extends RuntimeException {
+public class DataDeserializationException extends VeridotException {
     public DataDeserializationException(String message) {
         super(message);
     }